feat: add crud
