+++ .gitignore
@@ -19,3 +19,5 @@ yarn-error.log
 storage/redis/appendonly.aof
 storage/redis/dump.rdb
 /public/*
+.editorconfig
+/storage/src/*
