+++ src/Models/Meetup.php
@@ -7,7 +7,6 @@ namespace Blumilk\Meetup\Core\Models;
 use Blumilk\Meetup\Core\Formats;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class Meetup extends Model
 {
@@ -29,10 +28,10 @@ class Meetup extends Model
     {
         return $this->belongsTo(User::class);
     }
-    
-    public function speakers(): HasMany
+
+    public function speakers(): BelongsTo
     {
-        return $this->hasMany(Speaker::class);
+        return $this->belongsTo(Speaker::class);
     }
 
     public function organization(): BelongsTo
