+++ src/Providers/StoreFileProvider.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Blumilk\Meetup\Core\Providers;
+
+use Blumilk\Meetup\Core\Contracts\StoreFile;
+use Blumilk\Meetup\Core\Services\StoreFileService as Service;
+use Illuminate\Support\ServiceProvider;
+
+class StoreFileProvider extends ServiceProvider
+{
+    public function register(): void
+    {
+        app()->bind(StoreFile::class, fn(): Service => new Service());
+    }
+}
